+++ pkg/balancer/random/random.go
@@ -36,7 +36,7 @@ func (b *RandomBalancer) Proxies() []proxy.Proxy {
 
 func (b *RandomBalancer) Select(ctx context.Context, hzCtx *app.RequestContext) (proxy.Proxy, error) {
 	if b.proxies == nil {
-		return nil, balancer.ErrNoAvailable
+		return nil, balancer.ErrNotAvailable
 	}
 
 	if len(b.proxies) == 1 {
@@ -44,7 +44,7 @@ func (b *RandomBalancer) Select(ctx context.Context, hzCtx *app.RequestContext) 
 		if proxy.IsAvailable() {
 			return proxy, nil
 		}
-		return nil, balancer.ErrNoAvailable
+		return nil, balancer.ErrNotAvailable
 	}
 
 	failedReconds := map[string]bool{}
@@ -56,7 +56,7 @@ findLoop:
 	}
 	// no live upstream
 	if len(failedReconds) == len(b.proxies) {
-		return nil, balancer.ErrNoAvailable
+		return nil, balancer.ErrNotAvailable
 	}
 	failedReconds[proxy.ID()] = true
 	goto findLoop