+++ pkg/middleware/parallel/parallel.go
@@ -4,12 +4,13 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"sync"
+
 	"github.com/cloudwego/hertz/pkg/app"
 	"github.com/go-viper/mapstructure/v2"
 	"github.com/nite-coder/bifrost/internal/pkg/safety"
 	"github.com/nite-coder/bifrost/pkg/config"
 	"github.com/nite-coder/bifrost/pkg/middleware"
-	"sync"
 )
 
 type ParallelMiddleware struct {
@@ -47,7 +48,7 @@ func (m *ParallelMiddleware) ServeHTTP(ctx context.Context, c *app.RequestContex
 	}
 }
 func init() {
-	_ = middleware.RegisterMiddleware("parallel", func(params any) (app.HandlerFunc, error) {
+	_ = middleware.Register([]string{"parallel"}, func(params any) (app.HandlerFunc, error) {
 		if params == nil {
 			return nil, errors.New("parallel middleware params is empty or invalid")
 		}
@@ -62,7 +63,7 @@ func init() {
 		}
 		options := make([]*Options, 0)
 		for _, middlewareOption := range middlewareOptions {
-			h := middleware.FindHandlerByType(middlewareOption.Type)
+			h := middleware.Factory(middlewareOption.Type)
 			m, err := h(middlewareOption.Params)
 			if err != nil {
 				return nil, fmt.Errorf("%s middleware params is invalid in parallel middleware, error: %w", middlewareOption.Type, err)